+++ changelog.md
@@ -6,6 +6,8 @@
 2. Added MFA verification on user deletion (self-operation)  
 3. Deduplicated roles and event attendees on update  
 4. Prevented disabling password login when there is no other login method  
+5. Added audit log for updating global note  
+6. Removed the need to login when using email confirmation link  
 
 ## v2.8.2